refactor(events): move event initialization logic into initEvents method
