+++ src/init.test.ts
@@ -1,8 +1,8 @@
-import { Client, Intents, Message } from 'discord.js'
+import { Client, Intents } from 'discord.js'
 import * as initCommandsModule from './commands/init'
+import * as initEventsModule from './events/init'
 import { init } from './init'
 import { BotCommand, BotFunction, InitOptions } from './types'
-import * as handleEvent from './utils/handleEvent'
 
 const onceSpy = jest.fn()
 const loginSpy = jest.fn()
@@ -11,9 +11,7 @@ const mockNotLoggedInClient = {
   login: loginSpy,
 } as unknown as Client<false>
 
-const onSpy = jest.fn()
 const mockLoggedInClient = {
-  on: onSpy,
   application: {
     id: 'mock-application-id',
   },
@@ -35,10 +33,6 @@ describe('init', () => {
 
   const mockToken = 'mock-token'
 
-  const mockMsg = {
-    content: 'ping',
-  } as Message
-
   const mockCommands: BotCommand[] = [
     {
       name: 'ping',
@@ -58,12 +52,12 @@ describe('init', () => {
     token: mockToken,
   }
 
-  const handleEventSpy = jest
-    .spyOn(handleEvent, 'default')
-    .mockResolvedValue(undefined)
-
   const initCommandsSpy = jest
     .spyOn(initCommandsModule, 'default')
+    .mockReturnValue(undefined)
+
+  const initEventsSpy = jest
+    .spyOn(initEventsModule, 'default')
     .mockReturnValue(undefined)
 
   const setupTest = (options?: Partial<InitOptions>) =>
@@ -88,12 +82,6 @@ describe('init', () => {
     expect(Client).toHaveBeenCalledWith({
       intents,
     })
-  })
-
-  it('should subscribe to message events', async () => {
-    await setupTest()
-
-    expect(onSpy).toHaveBeenCalledWith('messageCreate', expect.any(Function))
   })
 
   it('should call initCommands with the given commands', async () => {
@@ -124,86 +112,24 @@ describe('init', () => {
     })
   })
 
+  it('should call initEvents with the given functions', async () => {
+    await setupTest()
+
+    expect(initEventsSpy).toHaveBeenCalledWith({
+      client: mockLoggedInClient,
+      functions: [mockFunction],
+      context: {
+        client: mockLoggedInClient,
+        functions: mockOptions.functions,
+      },
+    })
+  })
+
   it('should return a logged-in client', async () => {
     const client = await setupTest()
 
     expect(client).toStrictEqual(mockLoggedInClient)
     expect(loginSpy).toHaveBeenCalledWith(mockToken)
     expect(onceSpy).toHaveBeenCalledWith('ready', expect.any(Function))
-  })
-
-  it('should call handleEvent when an appropriate event was received', async () => {
-    await setupTest()
-
-    const eventHandler = onSpy.mock.calls[0][1] as (msg: Message) => void
-
-    eventHandler(mockMsg)
-
-    expect(handleEventSpy).toHaveBeenCalledWith([mockMsg], [mockFunction], {
-      client: mockLoggedInClient,
-      functions: mockOptions.functions,
-    })
-  })
-
-  it('should handle many functions and map each one to an event handler', async () => {
-    // A function without an explicit event will map to a messageCreate handler
-    const fnA: BotFunction = {
-      condition: jest.fn(),
-      callback: jest.fn(),
-    }
-
-    const fnB: BotFunction<'channelCreate'> = {
-      event: 'channelCreate',
-      condition: jest.fn(),
-      callback: jest.fn(),
-    }
-
-    const fnC: BotFunction<'messageCreate'> = {
-      event: 'messageCreate',
-      condition: jest.fn(),
-      callback: jest.fn(),
-    }
-
-    const fnD: BotFunction<'messageDelete'> = {
-      event: 'messageDelete',
-      condition: jest.fn(),
-      callback: jest.fn(),
-    }
-
-    const fnE: BotFunction<'channelCreate'> = {
-      event: 'channelCreate',
-      condition: jest.fn(),
-      callback: jest.fn(),
-    }
-
-    const mockFunctions = [fnA, fnB, fnC, fnD, fnE]
-
-    await setupTest({ functions: mockFunctions })
-
-    const expectedEvents = {
-      messageCreate: [fnA, fnC],
-      channelCreate: [fnB, fnE],
-      messageDelete: [fnD],
-    }
-
-    Object.entries(expectedEvents).forEach(([event, expectedFns], i) => {
-      expect(onSpy).toHaveBeenNthCalledWith(i + 1, event, expect.any(Function))
-
-      const eventHandler = onSpy.mock.calls[i][1] as (msg: Message) => void
-
-      eventHandler(mockMsg)
-
-      expect(handleEventSpy).toHaveBeenNthCalledWith(
-        i + 1,
-        [mockMsg],
-        expectedFns,
-        {
-          client: mockLoggedInClient,
-          functions: mockFunctions,
-        },
-      )
-    })
-
-    expect(handleEventSpy.mock.calls).toMatchSnapshot()
   })
 })
